Fix ReconciliationManagerTests Also create and move tests that are the responsibility of the ReconciliationBuilder.
